requirements.txt
fastapi~=0.103.2
tortoise-orm~=0.20.0
bcrypt~=4.0.1
aerich~=0.7.2